requirements.txt
pyqt5
pyqtdarktheme
pyyaml